updates to parameterize demo home and solr home

+++ doc_crawler_stack/package/scripts/setup.sh
@@ -10,7 +10,8 @@ DEMO_ROOT=$3
 #Logfile e.g. /var/log/doc-crawler.log
 LOGFILE=$4
 
-if [ ! -d "$SOLR_PATH" ]; then
+if [ ! -d "$SOLR_PATH" ]
+then
     #solr is not on 2.2 but is installed on sandbox 
 	adduser solr
 	mkdir $SOLR_PATH
@@ -26,14 +27,21 @@ if [ ! -d "$SOLR_PATH" ]; then
 	ln -s solr-4.7.2 solr
 fi
 
-sudo -u hdfs hdfs dfs -rmr /user/solr/data/rfi_raw
+sudo -u hdfs hdfs dfs -test -d /user/solr/data/rfi_raw
+if [ $? -eq 0 ]
+then
+	sudo -u hdfs hdfs dfs -rmr /user/solr/data/rfi_raw
+fi
 sudo -u hdfs hdfs dfs -mkdir -p /user/solr/data/rfi_raw
 sudo -u hdfs hdfs dfs -chown solr /user/solr
 sudo -u hdfs hdfs dfs -chmod -R 777 /user
 
 #Move search docs to HDFS
 cd $DEMO_ROOT
-rm -rf RFIsForSolr
+if [ -d RFIsForSolr ]
+then
+	rm -rf RFIsForSolr
+fi
 unzip RFIsForSolr.zip
 cd RFIsForSolr
